+++ eth/api_backend.go
@@ -329,6 +329,14 @@ func (b *EthAPIBackend) RPCGasCap() uint64 {
 	return b.eth.config.RPCGasCap
 }
 
+func (b *EthAPIBackend) GasBuffer() float64 {
+	buffer := b.eth.config.GasBuffer
+	if buffer == 0 {
+		return 1.0
+	}
+	return buffer
+}
+
 func (b *EthAPIBackend) RPCEVMTimeout() time.Duration {
 	return b.eth.config.RPCEVMTimeout
 }